% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_function.R
\name{bin.trajectory}
\alias{bin.trajectory}
\title{Cell aligner, binning and summary statistics function on trajectory}
\usage{
bin.trajectory(x, trajectory, n = NULL, stat = "median", qu1 = NULL,
  qu2 = NULL, ...)
}
\arguments{
\item{x}{a data frame of expression values to be summarized statictically along the trajectory. It does not contain trajectory variable.
It may contain factors (e.g. cell labels, clusters).}

\item{trajectory}{a vector of trajectory distances to be used for cell alignement and binning. This vector can be extracted
from the tspace_matrix object, a matrix of trajectory space distances. Trajectory vector has to correspond to cells in the data frame x.}

\item{n}{an integer specifying number of the bins along the trajectory, the higher the number the fewer cells within the bin.
If n = NULL, function will bin all the cells with the same distance into common bin. For this feature it may be useful to round decimal
values in trajectory vector.}

\item{stat}{a string defining to calculate the first derivative ('slope'), median ('median') or
arithmetic ('mean') mean statistics. The default is median.}

\item{qu1}{a percentage defining the quantile which will be the right border of the first bin. Default is NULL.}

\item{qu2}{a percentage defining the quantile which will be the left border of the last bin. Default is NULL.}
}
\value{
This function returns data frame in which cells are binned along the trajectory and expressions are
summarized by selected statistics.
}
\description{
Post trajectory analysis function for calcualting statistics on expression values along the isolated trajectory
}
