% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_lknn}
\alias{find_lknn}
\title{Helper function: find_lknn}
\usage{
find_lknn(knn, l, core_n)
}
\value{
l-kNN graph
}
\description{
find_lknn is a helper function to reduce a principal graph to l-kNN graph
This function works as in MATLAB, Note: try to optimize code for speed
}
\keyword{internal}
