% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{graphfinder}
\alias{graphfinder}
\title{Helper function: graphfinder}
\usage{
graphfinder(x, k, distance, core_n)
}
\value{
kNN graph
}
\description{
graphfinder is a helper function to find a principal graph using KernelKnn
function as a base function
}
\keyword{internal}
