% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExplorer.R
\name{runExplorer}
\alias{runExplorer}
\title{runExplorer}
\usage{
runExplorer()
}
\description{
Starts interactive UI for data exploration. It allows upload of a custom
csv file, up to 1.2 Gb, however large data sets will load slowly and
all changes within application will be slow. It includes T cell FACS data from
the pre-print (doi: https://doi.org/10.1101/336313)
}
