% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSpace.R
\name{tSpace}
\alias{tSpace}
\title{tSpace}
\usage{
tSpace(df, K = 20, L = 15, D = "pearson_correlation", graph = 5,
  trajectories = 100, wp = 20, ground_truth = F,
  weights = "exponential", dr = "pca", seed = NULL, core_no = 1)
}
\arguments{
\item{df}{a data frame or a matrix of expression values, which contain information on data straucture,
e.g. expression of variable genes, developmentally relevant genes/proteins}

\item{K}{an integer specifying the K-nearest-neighbors}

\item{L}{an integer specifying the random L out of K-nearest-neighbors, L < K, usually K = (4*L)/3}

\item{D}{a string specfying metric for distance calculation. Supported: ’euclidean’, ’pearson_correlation’,
’manhattan’, ’chebyshev’, ’canberra’, ’braycurtis’,  ’simple_matching_coefficient’, ’minkowski’,
’hamming’, ’mahalanobis’, ’jaccard_coefficient’, ’Rao_coefficient’}

\item{graph}{an integer specifying how many L-K-NN graphs will be used for final trajectory calculation}

\item{trajectories}{an integer specifying how many trajectories will be calculatedl. Default value is 100,
see ground_truth for more details}

\item{wp}{an integer specifying the number of waypoints for trajectory refinement}

\item{ground_truth}{a booolean (TRUE or FALSE) specifying if trajectories are calculated for every data point.
As default tSpace calculates an aproximation using 100 trajectories, which is usually sufficient for understanding
of developmental relations in single cell data. If set to TRUE, calculation time will be longer
and trajectories parameter will be overridden.}

\item{weights}{a string specfying method to calculate the weights for refinement of the trajectory distances. Supported:
uniform, linear, quadratic and exponential.}

\item{dr}{a string specifying type of embbeding for visualization. Options: 'pca', 'tsne' or 'both'.
'pca' embbeds trajectory space matrix in principal components,
'tsne' uses Rtsne function with parameters: perplexity = 25, max_iter = 2000, for details see documentation of Rtsne package,
'both' calculates pca and tsne}

\item{seed}{an integer specifying seed for set.seed function in order to have reproducible tsne.}

\item{core_no}{and integer specifying number of cores for parallelization, check how many cores your machine has and adjust accordingly}
}
\value{
tSpace returns a list of objects: 1. a matrix of pca and/or t-SNE embbedings of trajectory space matrix and input data,
2. pca and/or tsne objects. pca object contians all the outputs of pca analysis,
tsne contians all the outputs of the t-SNE analysis, see \code{\link[Rtsne:Rtsne]{Rtsne}}
}
\description{
tSpace is the main function for trajectory analysis. The algorithm is described in the publication
Dermadi et al. 2018
Originally, it was developed for single cell analysis, however it can be applied on any type of large data.
}
\details{
If you use it please cite:
}
